Move utils to smart-tools
