+++ web_ui/packages/smart-tools/index.ts
@@ -11,6 +11,9 @@ export { buildInferenceImageInstance, InferenceImage } from './src/inference-ima
 export { buildSSIMInstance, SSIM } from './src/ssim/ssim';
 export { type RunSSIMProps, type SSIMMatch } from './src/ssim/interfaces';
 
+export { buildGrabcutInstance, Grabcut } from './src/grabcut/grabcut';
+export { type GrabcutData } from './src/grabcut/interfaces';
+
 export {
     formatContourToPoints,
     approximateShape,
@@ -19,4 +22,6 @@ export {
     concatFloat32Arrays,
     stackPlanes,
     isPolygonValid,
+    getPointsFromMat,
+    getMatFromPoints,
 } from './src/utils/tool-utils';
